moved event.kt to a new folder
